REFACTOR: move all node-related declarations to nodes.go.
