--- src/empty_test.go
@@ -1,14 +1,0 @@
-package mpt
-
-import (
-	"testing"
-
-	"github.com/ethereum/go-ethereum/rlp"
-	"github.com/stretchr/testify/require"
-)
-
-func TestEmptyNodeHash(t *testing.T) {
-	emptyRLP, err := rlp.EncodeToBytes(EmptyNodeRaw)
-	require.NoError(t, err)
-	require.Equal(t, EmptyNodeHash, Keccak256(emptyRLP))
-}
