+++ src/nodes.go
@@ -1,8 +1,25 @@
 package mpt
 
 import (
+	"encoding/hex"
+	"fmt"
+
+	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/rlp"
 )
+
+////////////
+// Globals
+////////////
+
+var (
+	EmptyNodeRaw     = []byte{}
+	EmptyNodeHash, _ = hex.DecodeString("56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421")
+)
+
+////////////////////////////
+// Node-general definitions
+////////////////////////////
 
 type Node interface {
 	Hash() []byte // common.Hash
@@ -125,4 +142,159 @@ func Deserialize(serializedNode []byte, db DB) Node {
 	}
 
 	return FromRaw(rawNode, db)
+}
+
+//////////////////////////
+// Empty node definitions
+//////////////////////////
+
+func IsEmptyNode(node Node) bool {
+	return node == nil
+}
+
+///////////////////////////
+// Branch node definitions
+///////////////////////////
+
+type BranchNode struct {
+	Branches [16]Node
+	Value    []byte
+}
+
+func NewBranchNode() *BranchNode {
+	return &BranchNode{
+		Branches: [16]Node{},
+	}
+}
+
+func (b BranchNode) Hash() []byte {
+	return crypto.Keccak256(b.Serialize())
+}
+
+func (b *BranchNode) SetBranch(nibble Nibble, node Node) {
+	b.Branches[int(nibble)] = node
+}
+
+func (b *BranchNode) RemoveBranch(nibble Nibble) {
+	b.Branches[int(nibble)] = nil
+}
+
+func (b *BranchNode) SetValue(value []byte) {
+	b.Value = value
+}
+
+func (b *BranchNode) RemoveValue() {
+	b.Value = nil
+}
+
+func (b BranchNode) Raw() []interface{} {
+	hashes := make([]interface{}, 17)
+	for i := 0; i < 16; i++ {
+		if b.Branches[i] == nil {
+			hashes[i] = EmptyNodeRaw
+		} else {
+			node := b.Branches[i]
+			if len(Serialize(node)) >= 32 {
+				hashes[i] = node.Hash()
+			} else {
+				// if node can be serialized to less than 32 bits, then
+				// use Serialized directly.
+				// it has to be ">=", rather than ">",
+				// so that when deserialized, the content can be distinguished
+				// by length
+				hashes[i] = node.Raw()
+			}
+		}
+	}
+
+	hashes[16] = b.Value
+	return hashes
+}
+
+func (b BranchNode) Serialize() []byte {
+	return Serialize(b)
+}
+
+func (b BranchNode) HasValue() bool {
+	return b.Value != nil
+}
+
+///////////////////////////////
+// Extension node definitions
+///////////////////////////////
+
+type ExtensionNode struct {
+	Path []Nibble
+	Next Node
+}
+
+func NewExtensionNode(nibbles []Nibble, next Node) *ExtensionNode {
+	return &ExtensionNode{
+		Path: nibbles,
+		Next: next,
+	}
+}
+
+func (e ExtensionNode) Hash() []byte {
+	return crypto.Keccak256(e.Serialize())
+}
+
+func (e ExtensionNode) Raw() []interface{} {
+	hashes := make([]interface{}, 2)
+	hashes[0] = NibblesToBytes(AppendPrefixToNibbles(e.Path, false))
+	if len(Serialize(e.Next)) >= 32 {
+		hashes[1] = e.Next.Hash()
+	} else {
+		hashes[1] = e.Next.Raw()
+	}
+	return hashes
+}
+
+func (e ExtensionNode) Serialize() []byte {
+	return Serialize(e)
+}
+
+//////////////////////////
+// Leaf node definitions
+//////////////////////////
+
+type LeafNode struct {
+	Path  []Nibble
+	Value []byte
+}
+
+// TODO [Alice]: Marked for deletion.
+func NewLeafNodeFromNibbleBytes(nibbles []byte, value []byte) (*LeafNode, error) {
+	ns, err := FromNibbleBytes(nibbles)
+	if err != nil {
+		return nil, fmt.Errorf("could not leaf node from nibbles: %w", err)
+	}
+
+	return NewLeafNodeFromNibbles(ns, value), nil
+}
+
+func NewLeafNodeFromNibbles(nibbles []Nibble, value []byte) *LeafNode {
+	return &LeafNode{
+		Path:  nibbles,
+		Value: value,
+	}
+}
+
+// TODO [Alice]: Marked for deletion.
+func NewLeafNodeFromBytes(key, value []byte) *LeafNode {
+	return NewLeafNodeFromNibbles(NibblesFromBytes(key), value)
+}
+
+func (l LeafNode) Hash() []byte {
+	return crypto.Keccak256(l.Serialize())
+}
+
+func (l LeafNode) Raw() []interface{} {
+	path := NibblesToBytes(AppendPrefixToNibbles(l.Path, true))
+	raw := []interface{}{path, l.Value}
+	return raw
+}
+
+func (l LeafNode) Serialize() []byte {
+	return Serialize(l)
 }
